viele notices gefixt. starten geht jetzt mit nur 50 notices in der ersten loop
