+++ logs/server_2017_12_20.txt
@@ -1,0 +1,1 @@
+Scream Labs PHP BitTorrent Announce Socket Server 1.1 [Build: 20110504] (Beta)SERVER (20.12.2017 09:05:10): Initializing...SERVER (20.12.2017 09:05:10): Initialization OK!SERVER (20.12.2017 09:05:10): Server running on localhost:81 ...SERVER (20.12.2017 09:05:10): Cache -> 404.html (278 Bytes)SERVER (20.12.2017 09:05:10): Cache -> index.html (175 Bytes)
